Strip Out Backspace Characters
Strip out backspace characters from command line arguments. Allows for
edits to filenames, filepaths, command names etc. with backspace. What
you see on the screen is what will be submitted to the command line
program.
Tested by poorly typing and then correcting filepaths with correct
navigation and file printing.

Create fs_terminal_commands Module

Refactor cmd_line.c to move file system terminal commands into its own
module.
Tested by navigating directories and printing files successfully.

Documentation and Style Changes

Prevent Buffer Overflow

Catch when absolute path grows too long (past BUFFER_SIZE of 256). When
it gets too long, it halts and returns an error.
Tested by moving into a directory making the current working directory
~20 characters long and then trying to cd into a path 240 characters
long.

Remove Unneccessary Include, Prevent Buffer Overflow

Restrict Scope of Terminal Buffer

Move terminal buffer out of .h to restrict scope.
